Fix clang errors (#79)
* distribute reward logic with sysvar

* new bank update for epoch

* new bank update for epoch

* size up

* resolve segfault

* downsizing to pass tests

* more initialization

* clean up

* small bug fix

* adjust comments

* more structure

* remove header

* wip fix for fd_runtime_init_bank_from_genesis

* need to fix stake_history

* fix block height

* stake history sysvar fix

* remove unneeded header

* get rid of fd_stakes_init

* resolve seg faults + lamports getting closer

* resolve clang issue

* more fixes + logging

* fix clang errors

+++ src/flamenco/runtime/fd_account.h
@@ -90,7 +90,7 @@ static inline
 int fd_account_is_writable_idx( fd_txn_t const * txn_descriptor,
                                 uchar program_id,
                                 int idx ) {
-  ushort acct_addr_cnt = txn_descriptor->acct_addr_cnt;
+  int acct_addr_cnt = txn_descriptor->acct_addr_cnt;
   if (txn_descriptor->transaction_version == FD_TXN_V0) {
     acct_addr_cnt += txn_descriptor->addr_table_adtl_cnt;
   }
